requirements.txt
asgiref==3.11.1
Django==6.0.4
djangorestframework==3.17.1
gunicorn==25.3.0
packaging==26.1
pillow==12.2.0
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
python-decouple==3.8
six==1.17.0
sqlparse==0.5.5
tzdata==2026.1
whitenoise==6.12.0
